+++ src/app/_component/BlockPage/style.module.css
@@ -1,0 +1,12 @@
+.block-page {
+  width: 100%;
+  height: 100%;
+  background-color: rgba(0, 0, 0, 0.2);
+  color: aliceblue;
+  font-size: 4rem;
+  position: absolute;
+  top: 0;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
