+++ src/app/menu/layout.module.css
@@ -1,0 +1,8 @@
+/** name section */
+.name_section {
+  display: flex;
+  justify-content: space-between;
+  padding: 1rem 2rem;
+  height: var(--name-section-height);
+  background-color: beige;
+}
